:recycle: Columnクラスをcli.rbからcolumn.rbに移動しました
