+++ lib/object_pascal_analyzer/cli.rb
@@ -1,26 +1,13 @@
 require "object_pascal_analyzer"
 
+require "object_pascal_analyzer/column"
+
 require 'csv'
 require 'json'
 require 'thor'
 
 module ObjectPascalAnalyzer
   class Cli < Thor
-
-    class Column
-      attr_reader :key, :alignment, :title
-      def initialize(key, alignment, title)
-        @key, @alignment, @title = key, alignment, title
-      end
-
-      def row_format(max_length)
-        "%#{alignment}#{max_length}\{#{key.to_s}\}"
-      end
-
-      def header_format(max_length)
-        "%-#{max_length}\{#{key.to_s}\}"
-      end
-    end
 
     COLS = [
       Column.new(:path,          '-', 'Path'),
